Fixed bugs in opener event code
The selector for openers was selecting the LI element instaed of the actual
opener anchor, so the data wasn't being passed. We also learned that a
synchronous XHR request can block until it's finished before letting the
browser move on to another page, so we revert that part of the code.

Signed-off-by: Branko Vukelic <branko@outernet.is>